MHV-48182: Use alerts hook created and implemented
